Rtk query

+++ src/pages/Users/Users.jsx
@@ -3,6 +3,7 @@ import * as ReactDOM from "react-dom";
 import { Grid, GridColumn } from "@progress/kendo-react-grid";
 import users from "./Users.json";
 import './Users.css';
+import { useGetAllUsersQuery, useEditUserMutation, useCreateUserMutation } from "../../features/apiSlice";
 import { Button } from "@progress/kendo-react-buttons";
 import { Window } from "@progress/kendo-react-dialogs";
 import { uid } from 'uid';
@@ -46,22 +47,39 @@ const Users = () => {
   };
 
   
-  
+  const { data, error, isLoading, refetch } = useGetAllUsersQuery();
+  const [edit] = useEditUserMutation();
+  const [_addUser] = useCreateUserMutation();
   const [visible, setVisible] = React.useState(false);
   const [id, setId] = React.useState(null);
   // const [info, setInfo] = React.useState(users);
-  const info = useSelector((state) => state.info.users);
+  // const info = useSelector((state) => state.info.users);
+  // React.useEffect(() =>{
+  //   console.log(info)
+  // },[info])
+
+  React.useEffect(() =>{
+    console.log("data ", data)
+  },[data])
+
   const [formData, setFormData] = React.useState({});
+
+  const getById = id => {
+    const element = data.find(el => el.id === id);
+    return element;
+  }
+
   const openDialog = id => {
     console.log("Active");
     setVisible(1);
     setFormData({});
     setId(id);
-    setFormData({id, code: getCodeById(id), surname: getSurnameById(id)});
+    setFormData(getById(id));
   };
 
   const deleteUser = (id) => {
     //const arr = info.filter(el => el.id !== id);
+    if(window.confirm('Удалить пользователя?'))
     dispatch(removeUser(id));
     //setInfo(arr);
   }
@@ -69,41 +87,39 @@ const Users = () => {
   const closeDialog = () => {
     setVisible(0);
     setId(null);
-    setFormData({surname: '', code:''});
+    setFormData({surname: '', code:'', password:''});
   }
 
-  const getCodeById = id => {
-    console.log(info, id);
-    const element = info.find(el => el.id === id);
-    return element.code;
-  };
-  const getSurnameById = id => {
-    const element = info.find(el => el.id === id);
-    return element.surname;
-  };
+  // const getCodeById = id => {
+  //   const element = data.find(el => el.id === id);
+  //   return element.code;
+  // };
+  // const getSurnameById = id => {
+  //   const element = data.find(el => el.id === id);
+  //   return element.surname;
+  // };
   const save = () => {
     // const arr = info.map(el =>{
     //    if(el.id != id) return el;
     //    else return formData;
     // });
-    dispatch(editUser({id, formData}));
+    //dispatch(editUser({id, formData}));
+    edit(formData);
     //setInfo(arr);
     closeDialog();
   }
-  React.useEffect(() => {
-    console.log(formData)
-  },[formData])
   const addUser1 = () => {
     setVisible(2);
-    setFormData({surname: '', code:''});
+    setFormData({name:'', login:'', email:'', password:''});
   }
   const add = () => {
-    if(formData.code && formData.surname)
+    if(formData.name && formData.email && formData.login && formData.password)
     {
-      formData.id = uid();
+      //formData.id = uid();
       //setInfo([...info, formData]);
-      dispatch(addUser(formData));
-      setFormData({surname: '', code:''});
+      // dispatch(addUser(formData));
+      _addUser(formData)
+      setFormData({name:'', login:'', email:'', password:''});
       setVisible(0);
     }
   }
@@ -114,14 +130,17 @@ const Users = () => {
         <Button onClick={() => addUser1()}>Добавить</Button>
       </div>
     <Grid
-      data={info}
+      data={data}
       className="grid"
       style={{
         height: "400px",
       }}
     >
-      <GridColumn field="code" title="Code"  />
-      <GridColumn field="surname" title="Surname" />
+      <GridColumn field="name" title="Name"  />
+      <GridColumn field="email" title="Email"  />
+      <GridColumn field="login" title="Login"  />
+      {/* <GridColumn field="code" title="Code"  />
+      <GridColumn field="surname" title="Surname" /> */}
       <GridColumn cell={EditCell}  width="200px" />
       <GridColumn cell={DeleteCell}  width="200px" />
     </Grid>
@@ -132,13 +151,24 @@ const Users = () => {
             {visible === 1 ? <legend>User Details</legend> : <legend>Add User</legend>  }
 
             <label className="k-form-field">
-              <span>Code</span>
-              <input className="k-input" value={formData.code} onChange={(e) => setFormData({...formData, code: e.target.value})} placeholder="Code" />
+              <span>Name</span>
+              <input className="k-input" value={formData.name} onChange={(e) => setFormData({...formData, name: e.target.value})} placeholder="Name" />
             </label>
             <label className="k-form-field">
-              <span>Surname</span>
-              <input className="k-input" value={formData.surname} onChange={(e) => setFormData({...formData, surname: e.target.value})} placeholder="Surname" />
+              <span>Email</span>
+              <input className="k-input" value={formData.email} onChange={(e) => setFormData({...formData, email: e.target.value})} placeholder="Email" />
             </label>
+            <label className="k-form-field">
+              <span>Login</span>
+              <input className="k-input" value={formData.login} onChange={(e) => setFormData({...formData, login: e.target.value})} placeholder="Login" />
+            </label>
+            {
+              visible === 2 ? 
+            <label className="k-form-field">
+              <span>Password</span>
+              <input className="k-input" value={formData.password} type="password" onChange={(e) => setFormData({...formData, password: e.target.value})} placeholder="Password" />
+            </label>: <></>
+            }
           </fieldset>
 
           <div className="text-right">
